+++ src/components/metrics/card/index.js
@@ -1,24 +1,18 @@
 import React, {Component, PropTypes} from 'react'
 import ReactDOM from 'react-dom'
-import JSONTree from 'react-json-tree'
 
 import { connect } from 'react-redux';
 import { removeMetric, changeMetric } from 'gModules/metrics/actions.js'
 import { changeGuesstimate } from 'gModules/guesstimates/actions.js'
 import { changeGuesstimateForm } from 'gModules/guesstimate_form/actions.js'
 
-import Histogram from 'gComponents/simulations/histogram'
 import MetricModal from '../modal/index.js'
-import StatTable from 'gComponents/simulations/stat_table'
 import DistributionEditor from 'gComponents/distributions/editor/index.js'
-import DistributionSummary from 'gComponents/distributions/summary/index.js'
 import MetricToolTip from './tooltip.js'
-import MetricName from './name'
-import Icon from 'react-fa'
 import $ from 'jquery'
-import MetricToken from './token/index.js'
 import './style.css'
 import * as canvasStateProps from 'gModules/canvas_state/prop_type.js'
+import MetricCardViewSection from './ViewSection.js'
 
 const INTERMEDIATE = 'INTERMEDIATE'
 const OUTPUT = 'OUTPUT'
@@ -144,21 +138,6 @@ class MetricCard extends Component {
     $(this.refs.dom).focus();
   }
 
-  showSimulation() {
-    const stats = _.get(this.props, 'metric.simulation.stats')
-    if (stats && _.isFinite(stats.mean) && _.isFinite(stats.stdev) && _.isFinite(stats.length)) {
-      return (stats.stdev === 0 || (stats.length > 5))
-    } else {
-      return false
-    }
-  }
-
-  _shouldShowStatistics() {
-    const isScientific = (this.props.canvasState.metricCardView === 'scientific')
-    const isAvailable = this.showSimulation() && (_.get(this.props, 'metric.simulation.stats').length > 1)
-    return isScientific && isAvailable
-  }
-
   _focusForm() {
     const editorRef = _.get(this.refs, 'DistributionEditor.refs.wrappedInstance')
     editorRef && editorRef.focus()
@@ -170,11 +149,6 @@ class MetricCard extends Component {
     const {guesstimate} = metric
 
     const anotherFunctionSelected = ((metricClickMode === 'FUNCTION_INPUT_SELECT') && !isSelected)
-
-    const showSimulation = this.showSimulation()
-    const shouldShowStatistics = this._shouldShowStatistics()
-    const shouldShowJsonTree = (metricCardView === 'debugging')
-    const hasGuesstimateDescription = !_.isEmpty(guesstimate.description)
 
     const titleView = !this.props.hovered && !isSelected && this._isTitle()
     const relationshipClass = relationshipClasses[relationshipType(metric.edges)]
@@ -204,60 +178,15 @@ class MetricCard extends Component {
             onChange={this.handleChangeGuesstimate.bind(this)}
         />
 
-        <div className={`section ${metricCardView}`}>
-
-          {(metricCardView !== 'basic') && showSimulation &&
-            <Histogram height={(metricCardView === 'scientific') ? 110 : 30}
-                simulation={metric.simulation}
-                cutOffRatio={0.995}
-            />
-          }
-
-          <div className='row '>
-            <div className='col-xs-10 sqwish-right'>
-              <div className='row'>
-                {(!_.isEmpty(metric.name) || isSelected) &&
-                  <div className='col-xs-12'>
-                    <MetricName
-                      isSelected={isSelected}
-                      name={metric.name}
-                      onChange={this.handleChangeMetric.bind(this)}
-                      jumpSection={this._focusForm.bind(this)}
-                      ref='name'
-                    />
-                  </div>
-                }
-
-                <div className='col-xs-12'>
-                  {showSimulation &&
-                    <DistributionSummary
-                        guesstimateForm={guesstimateForm}
-                        simulation={metric.simulation}
-                    />
-                  }
-                </div>
-              </div>
-            </div>
-
-            <div className='col-xs-2 sqwish-middle'>
-              <MetricToken
-                 readableId={metric.readableId}
-                 anotherFunctionSelected={anotherFunctionSelected}
-                 onOpenModal={this.openModal.bind(this)}
-                 hasGuesstimateDescription={hasGuesstimateDescription}
-              />
-            </div>
-          </div>
-
-          {shouldShowJsonTree &&
-            <div className='row'> <div className='col-xs-12'> <JSONTree data={this.props}/> </div> </div>
-          }
-
-          {shouldShowStatistics &&
-            <div className='row'> <div className='col-xs-12'> <StatTable stats={metric.simulation.stats}/> </div> </div>
-          }
-
-        </div>
+      <MetricCardViewSection
+          metricCardView={metricCardView}
+          metric={metric}
+          isSelected={isSelected}
+          onChangeName={this.handleChangeMetric.bind(this)}
+          guesstimateForm={guesstimateForm}
+          anotherFunctionSelected={anotherFunctionSelected}
+          onOpenModal={this.openModal.bind(this)}
+        />
 
         {isSelected && !this.state.modalIsOpen &&
           <div className='section editing'>
